build: fixed some unboxing issues present to due a bug in the eclipse compiler for java 8. See Bug 440019

+++ plugins/de.cau.cs.kieler.klay.force/src/de/cau/cs/kieler/klay/force/model/EadesModel.java
@@ -44,8 +44,8 @@ public final class EadesModel extends AbstractForceModel {
     protected void initialize(final FGraph graph) {
         super.initialize(graph);
         maxIterations = graph.getProperty(Properties.ITERATIONS);
-        springLength = graph.getProperty(Properties.SPACING);
-        repulsionFactor = graph.getProperty(Properties.REPULSION);
+        springLength = graph.getProperty(Properties.SPACING).doubleValue();
+        repulsionFactor = graph.getProperty(Properties.REPULSION).doubleValue();
     }
     
     /**
