+++ plugins/de.cau.cs.kieler.klay.tree/src/de/cau/cs/kieler/klay/tree/p3place/NodePlacer.java
@@ -89,7 +89,7 @@ public class NodePlacer implements ILayoutPhase {
         progressMonitor.begin("Processor order nodes", 2);
 
         /** set the spacing according to the user inputs */
-        spacing = tGraph.getProperty(Properties.SPACING);
+        spacing = tGraph.getProperty(Properties.SPACING).doubleValue();
 
         /** find the root node of this component */
         LinkedList<TNode> roots = new LinkedList<TNode>();
